Extend the function to be able to receive conference feedback as well based on a flag in the request and save it in the ConferenceFeedback Entity
